requirements.txt
aksharamukha==2.0.1
alabaster==0.7.12
alembic==1.8.0
async-generator==1.10
attrs==21.4.0
Babel==2.10.3
backports.functools-lru-cache==1.6.4
bcrypt==3.2.2
black==22.3.0
cachetools==5.2.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
conllu==4.4.2
coverage==6.4.1
cryptography==37.0.2
Deprecated==1.2.13
dnspython==2.2.1
docutils==0.17.1
email-validator==1.2.1
fabric==2.7.0
Flask==2.1.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.1
Flask-WTF==1.0.1
fonttools==4.33.3
google-api-core==2.8.2
google-auth==2.9.0
google-cloud-vision==2.7.3
googleapis-common-protos==1.56.3
greenlet==1.1.2
grpcio==1.47.0
grpcio-status==1.47.0
gunicorn==20.1.0
h11==0.13.0
idna==3.3
imagesize==1.3.0
importlib-metadata==4.11.4
indic-transliteration==2.3.31
iniconfig==1.1.1
invoke==1.7.1
itsdangerous==2.1.2
jaconv==0.3
Jinja2==3.1.2
langcodes==3.3.0
language-data==1.1
lxml==4.9.0
Mako==1.2.0
marisa-trie==0.7.7
MarkupSafe==2.1.1
memory-profiler==0.60.0
mypy-extensions==0.4.3
outcome==1.1.0
packaging==21.3
paramiko==2.11.0
pathlib2==2.3.7.post1
pathspec==0.9.0
Pillow==9.1.1
platformdirs==2.5.2
pluggy==1.0.0
proto-plus==1.20.6
protobuf==3.20.1
psutil==5.9.1
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.12.0
pykakasi==2.2.1
PyMuPDF==1.20.1
PyNaCl==1.5.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.1.2
pytest-cov==3.0.0
pytest-cover==3.0.0
pytest-coverage==0.0
python-dotenv==0.20.0
python-slugify==6.1.2
pytz==2022.1
PyYAML==6.0
regex==2022.6.2
requests==2.27.1
roman==3.3
rsa==4.8
selenium==4.2.0
six==1.16.0
sniffio==1.2.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
Sphinx==5.0.2
sphinx-rtd-theme==1.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.4.37
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
trio==0.20.0
trio-websocket==0.9.2
typer==0.4.1
typing_extensions==4.2.0
unicodedata2==14.0.0
urllib3==1.26.9
Werkzeug==2.1.2
wrapt==1.14.1
wsproto==1.1.0
WTForms==3.0.1
zipp==3.8.0
